refactor(ssid): extract phase1 audit cluster + SiteRowInputs dataclass (phase 2/6)
Extract the read-only audit workflow from ssid_template_consolidation
into _ssid_template_phase1.py, mirroring the phase 1 cache-cluster
split. The parent module now holds only phase 2-5 orchestration plus
its own _fetch_and_log copy (kept in the parent namespace so tests
that patch ssid_template_consolidation.mistapi see the mock at call
time -- the function's __globals__ must match the patched module).

* New _ssid_template_phase1.py (601 LOC):
  - _fetch_and_log (module-local mistapi lookup for _fetch_all_org_data)
  - _build_mxtunnel_lookup / _build_template_lookup / _build_sitegroup_lookup
  - _resolve_template / _get_template_wlans / _find_target_wlan
  - _classify_site / _determine_target_group
  - _SiteRowInputs dataclass + _assemble_site_row (drops the 13-param
    signature that had tripped STRUCT-PARAMS)
  - _build_site_row / _collect_key_values / _collect_comparison_keys
  - _collect_group_wlan_configs / _analyze_group_deviations
  - _build_deviation_record / _group_by_target
  - _detect_cross_cluster_drift / _append_drift_record
  - _print_phase1_summary
  - _SsidTemplatePhase1Cluster (proxies via _ClusterBase._mm)

* Parent module drops ~537 LOC of phase-1 helpers, adds a 21-symbol
  re-export from _ssid_template_phase1, binds the new cluster into
  self._clusters, and defines its own _fetch_and_log so patch.object
  on _mod.mistapi is observed.

* Tests unchanged (241/241 pass); mypy --strict clean.
